unify pipeline model with PretrainModelPipe.
